pyproject.toml
[project]
name = "cv-jpsrocha"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepl>=1.22.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
]
